requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
concurrent-log-handler==0.9.20
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.13
django-ckeditor==6.4.2
django-cors-headers==3.7.0
django-debug-toolbar==3.7.0
django-hosts==5.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-xml==2.0.0
drf-yasg==1.20.0
gevent==21.12.0
greenlet==1.1.3.post0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==21.3
Pillow==9.1.1
portalocker==2.5.1
prompt-toolkit==3.0.31
psutil==5.9.2
pycparser==2.21
pyftpdlib==1.5.6
PyJWT==2.4.0
pyOpenSSL==19.1.0
pyparsing==3.0.9
pytz==2022.4
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.12
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.5.0
